+++ src/main/uni_hid_parser_ds4.c
@@ -421,20 +421,21 @@ void uni_hid_parser_ds4_set_led_color(uni_hid_device_t* d, uint8_t r, uint8_t g,
   uni_hid_device_send_intr_report(d, (uint8_t*)&ff, sizeof(ff));
 }
 
-void uni_hid_parser_ds4_set_rumble(uni_hid_device_t* d, uint8_t left,
-                                   uint8_t right, uint16_t duration) {
+void uni_hid_parser_ds4_set_rumble(uni_hid_device_t* d, uint8_t value,
+                                   uint8_t duration) {
+  // FIXME: timer that cancels rumble after duration
   UNUSED(duration);
 
-  // Left motor: big force
-  // Right motor: small force
   ds4_ff_report_t ff = {0};
 
   ff.transaction_type = 0xa2;     // DATA | TYPE_OUTPUT
   ff.report_id = 0x11;            // taken from HID descriptor
   ff.unk0[0] = 0xc4;              // HID alone + poll interval
   ff.flags = DS4_FF_FLAG_RUMBLE;  // blink + LED + motor
-  ff.rumble_left = left;
-  ff.rumble_right = right;
+  // Right motor: small force
+  // Left motor: big force
+  ff.rumble_right = value;
+  ff.rumble_left = value;
 
   /* CRC generation */
   uint8_t bthdr = 0xA2;