+++ src/main/uni_hid_parser_ds4.h
@@ -34,7 +34,7 @@ void uni_hid_parser_ds4_parse_raw(struct uni_hid_device_s *d,
                                   const uint8_t *report, uint16_t len);
 void uni_hid_parser_ds4_set_led_color(struct uni_hid_device_s *d, uint8_t r,
                                       uint8_t g, uint8_t b);
-void uni_hid_parser_ds4_set_rumble(struct uni_hid_device_s *d, uint8_t left,
-                                   uint8_t right, uint16_t duration);
+void uni_hid_parser_ds4_set_rumble(struct uni_hid_device_s *d, uint8_t value,
+                                   uint8_t duration);
 
 #endif  // UNI_HID_PARSER_DS4_H